requirements.txt
crewai
langchain
langchain-openai
python-dotenv
crewai[tools]
pypdf2